Rename `ICanBoogie\Core\BootEvent` as `ICanBoogie\Application\BootEvent`
