+++ tests/Application/BootEventTest.php
@@ -9,9 +9,8 @@
  * file that was distributed with this source code.
  */
 
-namespace ICanBoogie\Core;
+namespace ICanBoogie\Application;
 
-use ICanBoogie\Core;
 use ICanBoogie\Application;
 
 use function ICanBoogie\app;
@@ -29,7 +28,7 @@ class BootEventTest extends \PHPUnit_Framework_TestCase
 
 		$called = false;
 
-		app()->events->once(function (BootEvent $event, Core $target) use ($app, &$called) {
+		app()->events->once(function (BootEvent $event, Application $target) use ($app, &$called) {
 
 			$this->assertSame($app, $target);
 			$event->stop();
